refactor(activerecord): house IndexDefinitionRow next to IndexDefinition
